# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mianyu MVP HTTP API
  version: 1.0.0
  description: 五模块唯一 HTTP 契约。除匿名注册和健康检查外，请求均携带 X-User-Id；写请求携带 Idempotency-Key。
servers:
  - url: http://127.0.0.1:8787/api/v1
...
paths:
...
  /users/anonymous:
    post:
      tags: [Common]
      operationId: createAnonymousUser
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { age_mode: { $ref: '#/components/schemas/AgeMode' } } } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } } }
...
  /feedback:
    post:
      tags: [Module5-Growth]
      operationId: submitFeedback
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MorningFeedback' } } } }
      responses: { '201': { $ref: '#/components/responses/Success' }, '409': { $ref: '#/components/responses/Error' } }
...
  /archive:
    get:
      tags: [Module5-Growth]
      operationId: archive
      parameters: [{ name: period, in: query, required: true, schema: { type: string, enum: [7d, 30d] } }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
components:
  parameters:
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 100 } }
  responses:
    Success: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    Error: { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
    Envelope: { type: object, required: [data], properties: { data: {} } }
    ErrorEnvelope: { type: object, required: [error], properties: { error: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string }, details: {} } } } }
    AgeMode: { type: string, enum: [adult, child] }
    VoicePreference: { type: string, enum: [want, avoid, unspecified] }
...
    MorningFeedback:
      type: object
      additionalProperties: false
      required: [session_id, fall_asleep_ease, sound_comfort, voice_next_time]
      properties:
        session_id: { type: string }
        fall_asleep_ease: { type: string, enum: [easy, normal, difficult, unknown] }
        sound_comfort: { type: string, enum: [comfortable, acceptable, uncomfortable] }
        voice_next_time: { $ref: '#/components/schemas/VoicePreference' }
        story_effect_rating: { type: integer, minimum: 1, maximum: 5 }
        disliked_content: { type: string, maxLength: 500 }
        note: { type: string, maxLength: 500 }
        forbidden_sound_tags: { type: array, items: { type: string } }
        story_themes: { type: array, items: { type: string } }
        breath_templates: { type: array, items: { type: string } }
        scenes: { type: array, items: { type: string } }